demos/frequency-deployment/requirements.txt
azure-functions
azure-identity
pyodbc
requests
PyJWT[crypto]
cryptography
